Smaller changes to OPT3001 driver

+++ src/driver/opt3001_drv.h
@@ -34,10 +34,10 @@ static const uint16_t CONF_100MS_SINGLE = 0b1100001000010000;
 class opt3001_drv {
 public:
 	opt3001_drv(i2c_interface & i2c, uint8_t i2c_addr);
-	void  start_measure(uint16_t conf);
-	bool  measure_ready();
-	float get_light();
-	bool  detect_sensor();
+	void     start_measure(uint16_t conf);
+	bool     measure_ready();
+	uint32_t get_light();
+	bool     detect_sensor();
 
 private:
 	i2c_interface & _i2c;
